link-identities(fix): ensure non-empty delegations
It was discovered that if an update to a Person document was passed an
empty set of delegations, then the document would be updated where the
delegations were set to the empty list. This is a bug and should not be
possible to do.

This patch ensures that the Direct and Indirect delegations are nonempty
when constructing them. For Indirect this is achieved by checking in
`try_from_iter` if the indirect and direct are empty, failing if they
are. For Direct we change the inner type from BTreeSet to
NonEmptyOrderedSet. We remove the FromIterator implementation and
introduce a `try_from_iter` -- similar to Indirect's construction. We
also provide a way to create a Direct through `new` and the ability to
to add more keys via `insert` and `extend`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>