link-canonical: derive feature for serde
If one attempted to build link-canonical by itself, there would be a
compilation failure saying `Serialize` was not in scope. This was due to
the "derive" feature not being enabled. My guess is that it was
compiling with the rest of the suite since it had some serde available
to it that *had* enabled the feature.

Nonetheless, this patch adds the feature flag to get compilation working
again.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>