link-canonical: add canonical json
We add a Canonical JSON[0] compliant Value type to the link-canonical
package. The Value type consists of the usual Object, Array, String,
Number, Bool, and Null values. However, the Map and Array variants are
backed by a BTreeMap and BTreeSet, respectively. Strings use the Cstring
type defined in link-canonical. Numbers can only be u8..u64 and i8..u64.

There is a ToCjson trait which consists of a function that can serialize
the implementing type into a Value.

[0]: http://wiki.laptop.org/go/Canonical_JSON

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>