librad, link-identities, rad-identities: remove Role type
To avoid the prescription of naming (which is hard), we remove the Role
type. Instead of this, when a Peer is denoted as Replicated we simply
have a single generic parameter.

In librad::git::identities::relations, we introduce a new type Persona
which consists of a generic parameter -- generally fulfilled by Person,
but made generic to allow for mapping into othe types. It also holds an
optional `PeerId` which is the identifier that this peer is delegating
with (if any), and also the `Refs` they may have.

Note that this is enough information to calculate the information we
used before:
* Maintainer means that the `PeerId` is set
* Contributor means that the `Refs` is set
* Tracker means that neither were set

Changing this type required the `tracked` functions in rad-identities to
also be updated.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>