link-canonical, link-canonical-derive, test: rework derive
The derive logic was left in the wrong state, due to some git history
issues. A few things that were wrong:
* The struct output was made up of an object with the struct name as a
key and the value were the fields.
* The enum output was similar in that its variant names were used as
keys.
* Not part of the logic, but the formatting in quote was broken.

These problems were rectified by:
* The struct output only being the fields
* The enum output is either internally or adjacently tagged. The former
can be used with named or unit fields, but not with unnamed fields. The
latter can be used with all types of fields.
* rustfmt::skip::macros(quote) is used to skip formatting of the quote
macro.

Other hygienic changes were:
* rename_all_rule contained a `Err("TODO")`. This now provides a better
error message.
* There is now `FromIterator` impl for `Value` which allows one to
create a `Value::Object` using `collect`.
* The test cases are updated to reflect the new changes.
* Some organisation was made in the derive lib.rs.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>