librad: wire up default_only check
When the gossip protocol receives gossip, the urn and peer are checked
if they are tracked. We add an extra check to see if the operator has
tracked the urn with the default entry, and only that entry. If this is
the case, the `urn` is considered tracked and the replication will
occur.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>