link-tracking: fuse actions for same urn/peer pairs
This patch introduces the implementation of fusing tracking actions
when performing them in a batch fashion.

The main mechanism for this is the `Compose` trait in the submodule
`policy::compose`. It allows the composition of the `Track` and
`Untrack` policies. Unfortunately, since there is a single rule that
does not reduce, we introduce a `Reduction` type which captures this
stuckness. This becomes the main output for track and untrack
compositions.

The fusion of actions iterates over each action keeping track of them
in one of two maps -- splitting between a PeerId being present or not
-- and composing it if there was an existing entry. It turns these
back into an iterator of actions to continue the `batch` process.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>