link-tracking: optional prune on untracking
The v1 tracking implementation provided pruning when `untrack`
was called. While this might not always be the right thing to do, the
v2 implementation gives the option to prune when calling `untrack` and
`untrack_all`. This is provided as flag in the parameters. Due to the
flag and the new return type, this a backwards incompatible change.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>