node-lib: seed refactor and storage
This patch aims to clean up the seed code as well introduce storage
for seeds.

The refactoring rewrites the Seed struct to be generic over its
address. This allows us to reuse the struct in scenarios where the
Seed may contain a DNS address which later gets resolved to socket
addresses. It also aids in testing of seed related code.

The storage introduces two effects for storage: `Store` and
`Scan`. The `Store` trait provides `get`, `update`, `insert`, and
`remove` methods. The `Scan` trait provides `scan` for getting an
iterator of the stored seeds.

THe concrete storage implementation uses the `pickledb` package for a
lightweight solution to a key-value file for storing the seeds. They
are stored as JSON where each key is a PeerId and the value is the address.

Upon start up of the node, we load the seeds stored in
`Paths::seed_dir` as well as using the bootstrap seeds passed in from
the args.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>