node-lib: seed refactor and storage
This patch aims to clean up the seed code as well introduce storage
for seeds.

The refactoring rewrites the Seed struct to be generic over its
address. This allows us to reuse the struct in scenarios where the
Seed may contain a DNS address which later gets resolved to socket
addresses. It also aids in testing of seed related code.

The storage introduces a `Store` effect that provides `scan` for
getting an iterator of the stored seeds. The expected format of the
seeds is the form `<peer id>,<addr>[,<label>]`.

The concrete storage implementation uses a path that can be turned
into an iterator over the lines of the file, each line being parsed on
the call to `next`.

If no bootstrap nodes are provided upon the start up of the node, the
fallback option is to read the seeds config -- using the max active
parameter from the membership protocol settings.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>