ci: nixos sourcehut build
Setting up NixOS build for sourcehut, as well as having a reliable way
of developing using nix.

We use niv to pin our nix dependencies, that is niv itself, nixpkgs,
and rust-overlay. These dependencies are then imported in default.nix
and shell.nix which declare the build and development dependencies
respectively. The difference between the development deps are a few
nice tools like rust-anaylzer, cargo-expand, and ripgrep.

The NixOS build manifest then calls:

  nix-shell default.nix --run cmd

for each build script (or command).

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>