net: fix to allow streams above 23
It was discovered that the encoding of the stream types would cause an
extra byte to be writting in the buffer if the u8 value was above
23. This is due to CBOR encoding of integers and using an extra byte
to signify an integer larger than 23. This can be seen in the minicbor
implementation[0] of `Encoder::u8`.

This was not a problem in itself, the issue was that despite using a
fixed length array we used a call to end[1] which adds a CBOR break
byte. The previous implementation was lucky and consumed that
byte. With the addition of stream type 200 the break byte would be
left in the buffer causing decode errors down the line.

The fix, and to keep backwards-compatibility, is to apply the break
character to any stream in the range 0..=23 and to not apply it in the
range 0..=255.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>