peer: wire up request_pull
With the request_pull protocol implementation in place, we wire it up
to the peer along the tincan wires.

It is wired up similar to the interrogation protocol, however, since
it can respond with multiple progress messages before a success or an
error it must use a mpsc channel for consuming the stream of
responses. This meant a change was made to the `request_response` to
give `single_response` and `multi_response` functions instead of the
`request` function.

Smoke tests were added for request_pull as well as adding it to the
set of upgrade tests.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>