node-lib: tracking for peer, urn, or pair of both
This patch reworks the node-lib tracker to be able to check if
something is tracked for either a peer id, urn, or a pair of
both. This allows filtering to be more discriminate if an operator
does not want to hear updates from *all* peers for urn or *all* urns
for a peer.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>