rfc: amend RFC 702 wire format
The initial wire format did not line up with the style of existing
RPCs in librad. This patch redefines it to a request/response style
and removes the need for the type code.

It also preempts the introduction of pruning of references and adds a
new field to the success response.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>