node-lib: type directed request/response
The previous implementation of the request/response API only
implemented announcements, which in the case of responding did not
have any payload.

While introducing request-pull, we realised that a payload is
neccessary. This prompted us to have a type directed response enum,
allowing us to have a type safe way of making a request and expecting
the correct type of response back.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>