rfc: amend RFC-0696 with request-pull
This patch amends RFC 696 by introducing the request-pull RPC
method. It includes a short description description of the method and
includes the request and response payloads in the CBOR wire types.

It also changes the request-id to be an array after exploring some
refactoring (that is introduced in a later commit). We decided that
this would be a good idea in the case of code migrations.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>