paths: move seeds to seeds_file attribute
The seeds function was outright wrong. It would *always* use the
project directories, which would mean it was incompatible with the use
of Paths initialised with a root provided. As well as this, it would
not be per profile.

Make the Paths have a new field, `seeds_file`, which points to the
seeds file storage under the Paths home and Profile.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>