protocol: remove State param from recv::git
The recv::git portion of the protocol is required by the client that
will not take part in the gossip network. Thus, it is necessary to not
pass the `protocol::State` to the `git` function call.

Make the `git` function call only require the `Paths` parameter,
ie. removing `protocol::State`.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>