net: send only endpoint
Provide a send only quinn::Endpoint for clients that will only make
requests and do not require any incoming connections.

A ConnectPeer trait is also provided to capture the ability to connect
to a remote peer's socket. This can return just a Connection, where
the connection provider handles the incoming streams elsewhere, or it
can provide the Connection and the streams for handling locally.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>