peer: expose RPC methods via Client
The Peer and Client have similar functionality for the functions
`replicate`, `interrogate`, and `request_pull`.

Make the Peer implement these functions via the Client by adding an
`rpc_client` method that gives a Client via the Peer's Config. This
function is then used to call the respective functions as Client methods.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>