librad::replication: provide From<context::Urn> for Urn
When wanting to convert to from `context::Urn` to `Urn`, the only
current way to do so is through the Deref impl. However, this can be a
waste of a clone if an owned `context::Urn` is being converted.

Provide a `From<context::Urn>` for `Urn` to allow this owned conversion.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>