link-hooks
The implementation of RFC 703[0].

This crate is broken down into the following components:
* data and track -- for representing the URN and tracking changes that
get communicated to the individual hooks. They are kept generic over
their revision type and can be rendered to and parsed from
strings.
* hook -- the types and traits for running hooks and communicating to
them. A single `Hook` is represented by its path to the executable and
the subprocess it is running. This subprocess must implement the
`Handle` trait which knows how to `spawn` the process, `write` to the
process, and `wait_or_kill` the process. A set of `Hooks` is run using
an unordered manner, getting passed `Notification`s via a stream and
then terminated once the stream is exhausted.

[0]: https://github.com/radicle-dev/radicle-link/blob/master/docs/rfc/0703-storage-hooks.adoc

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>