link-hooks: smoke test running hooks
Add smoke tests for executing a number of hooks.

This is achieved by adding executables under `test/hooks`:
* `echo-data` - handles `Data` messages
* `echo-track` - handles `Track` messages
* `echo-forever` - acts a misbehaving hook and sleeps to simulate a
  hang

The tests pass file handles to the corresponding hook and ensure that
all the notifications were written to the file once the `Hooks`
process is finished.

In the case of using `echo-forever`, we are ensuring that a hanging
hook does not poison the `Hooks` process and all hooks complete their
respective tasks.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>