lnk-identities: make name optional for include remotes
If a remote peer does not include a `rad/self` then the include
generation would not include them as a git remote.

The use of `rad/self` is to simply supply the `name` so that remotes
can be more easily remembered and are human-friendly. This does not
mean that they are necessary.

Make `<name>@` optional for remotes added to include files, falling
back to `<peer id>` only as the remote name.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>