link-hooks: newtype and re-exports refactor
The config types for a Hook, buffer size & timeout duration, used the
bare types. This makes it harder to define a default for these
individual values outside of calling `config::Hook::default`. This
definition of a default is useful for arg/config parsers, for example,
when wiring up to lnkd.

Make two newtypes, Buffer & Timeout, that have the required trait
implementations for parsing and defaulting.

Also, to make things easier for importing, re-export Config from
librad::git::hooks.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>