lnk-identities: person/project delegates
A feature lacking from the `lnk identities` CLI is the ability to list
delegates of an identity. This *could* be derived from the output of
`tracked`. A dedicated subcommand is more convenient for this specific
output and is easier to remember.

Add a `delegates` subcommand to `person` and `project` that outputs
the delegates of a specific, verified identity. In the case of a
project it outputs an object with `"direct"` and `"indirect"` keys --
corresponding to `PeerId`s and `Urn`s respectively. In the case of a
person it outputs a list of `PeerId`s.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>